Rename DualQuaternion
