+++ s2e-ff/src/library/math/dual_quaternion.cpp
@@ -1,4 +1,4 @@
-#include "DualQuaternion.hpp"
+#include "dual_quaternion.hpp"
 
 #include <float.h>
