+++ s2e-ff/src/library/math/dual_quaternion.hpp
